feat: implement asset calculation and CSV reading functionality
