+++ src/main/java/com/app/financial/investmentassetapp/controller/AssetController.java
@@ -79,5 +79,4 @@ public class AssetController {
         return ResponseEntity.created(null).build();
     }
 
-
 }
